Simpify example of converting mlp model
