+++ border-core/src/dummy.rs
@@ -10,12 +10,12 @@ impl crate::Obs for DummyObs {
     }
 }
 
-// TODO: Consider to make this work with feature flag tch.
-// impl Into<tch::Tensor> for DummyObs {
-//     fn into(self) -> tch::Tensor {
-//         unimplemented!();
-//     }
-// }
+#[cfg(feature = "tch")]
+impl Into<tch::Tensor> for DummyObs {
+    fn into(self) -> tch::Tensor {
+        unimplemented!();
+    }
+}
 
 #[derive(Clone, Debug)]
 /// Dummy action.
@@ -27,30 +27,30 @@ impl crate::Act for DummyAct {
     }
 }
 
-// TODO: Consider to make this work with feature flag tch.
-// impl Into<tch::Tensor> for DummyAct {
-//     fn into(self) -> tch::Tensor {
-//         unimplemented!();
-//     }
-// }
+#[cfg(feature = "tch")]
+impl Into<tch::Tensor> for DummyAct {
+    fn into(self) -> tch::Tensor {
+        unimplemented!();
+    }
+}
 
-// TODO: Consider to make this work with feature flag tch.
-// impl From<tch::Tensor> for DummyAct {
-//     fn from(_value: tch::Tensor) -> Self {
-//         unimplemented!();
-//     }
-// }
+#[cfg(feature = "tch")]
+impl From<tch::Tensor> for DummyAct {
+    fn from(_value: tch::Tensor) -> Self {
+        unimplemented!();
+    }
+}
 
 #[derive(Clone)]
 /// Dummy inner batch.
 pub struct DummyInnerBatch;
 
-// TODO: Consider to make this work with feature flag tch.
-// impl Into<tch::Tensor> for DummyInnerBatch {
-//     fn into(self) -> tch::Tensor {
-//         unimplemented!();
-//     }
-// }
+#[cfg(feature = "tch")]
+impl Into<tch::Tensor> for DummyInnerBatch {
+    fn into(self) -> tch::Tensor {
+        unimplemented!();
+    }
+}
 
 /// Dummy batch.
 pub struct DummyBatch;
